requirements.txt
pyzmq
llama-index-core
llama-index-llms-groq
python-dotenv
torch
